+++ examples/main/variables.tf
@@ -1,0 +1,19 @@
+variable "client_name" {
+  description = "Client name"
+  type        = string
+}
+
+variable "environment" {
+  description = "Environment name"
+  type        = string
+}
+
+variable "stack" {
+  description = "Stack name"
+  type        = string
+}
+
+variable "azure_region" {
+  description = "Azure region to use"
+  type        = string
+}