+++ modules/automation-account/variables.tf
@@ -49,7 +49,7 @@ variable "custom_automation_account_name" {
 variable "log_analytics_resource_group_name" {
   description = "Log Analytics Workspace resource group name (if different from `resource_group_name` variable.)"
   type        = string
-  default     = ""
+  default     = null
 }
 
 variable "log_analytics_workspace_id" {